+++ rule/pom.xml
@@ -106,8 +106,7 @@
         <module>pulsar-client</module>
         <module>pulsar-functions</module>
         <module>dynamic</module>
-        <module>cxf-client</module>
-        <module>cxf-server</module>
+        <module>cxf</module>
       </modules>
       <build>
         <resources>
